requirements.txt
python-dotenv==1.0.1
requests==2.32.3
types-requests==2.32.0.20240622